Refactored session handling.
The session timeout is now solely handled by the build in servlet session. Our own `Session` object is persisted, and an `HttpSessionBindingListener` used to mark the persisted session as invalid when the http session invalidates.

The existing supplemental cookie has been completely removed.

The hard part of this was dealing with the auto-redirect that should happen when the server side session times-out. The `/app/verify` call would *always* keep the session alive, so it was impossible to test if the session was valid or not.

This was deal with by using a `TomcatContextCustomizer` that installs a custom `StandardManager` and news a custom `StandardSession` that has flags for preventing the access times being updated.

The mechanism for keep the session alive on large uploads has also been changed. It will change `HttpSession.setMaxInactiveInterval()` to no-timeout before the operation and set it back afterwards.