Some infrastructure to support rolling back of non-database changes. It was found that if there are operations during the setup wizard that are NOT database changes, then if there is a problem later on, that means there is a database rollback, those out-of-band operations are not undone.
This particularly shows itself on the VPN. It now does 3 main things that are outside of database transactions.

 * Setup a Wireguard interface
 * Start an SSH server
 * Integrate with VMSee.

If the last step (integration with VMSee) fails, then the low level wireguard interfaces  and the SSH server socket will be left running, but the database objects that supports these will be gone. This causes all kinds of oddities if you then try and correct the last step (e.g. wrong tenant) domain, and you will be effectively stuck in the setup wizard. Only a *reboot* will clear this.

This cannot really be dealt with on a case by case basis. There needs to be framework to allow other types of transaction behaviour.

To address this, I have added `TransactionService.tx()`. This `ThreadLocal` may be called anywhere inside a transaction, and has an `undoable()` method which may be passed a `Runnable` that is called if the transaction fails.

This is not perfect, as it still means the transaction is not undone in the exact same order as it should really be. Instead, the database changes are undone, and then any other `undoable()` actions are running. This may not be in the exact same order that they were run, but it should be good enough to support the setup wizard case.

Longer term we really need a better transaction API.