Improvements to home page resolution. This is now implemented as a interface `HomePageResolver` instead of an annotation `HomePage`. I needed to make decisions what home page to use dynamically. Apart from the use of the annotation mean that it was static, and only the presence or not of an extension determining what the home page would be.
The second problem was that the home page was always set up upon `AuthenticationState` creation. This meant it could and would happen before any kind of login. The same home page would then persist right through to the session itself. Instead, no home page is now set by default on the state. If the home page is null, the caller is expected to resolve the default. There is a new helper method in `AuthenticationService.getCurrentPage()` that should be used where appropriate.