Following a late night discussion with @ludup, and discovering more bugs, I attempted to clean up some of that redirecting code.
This is the result. The general idea was to separate all of the calls to `getCurrentPage()`, both direct and indirect, into whether they only really want to know what the current page, and those that are expecting a redirect or the state to be reset. This was actually pretty clear in the end, and the usages were probably split evenly.

The calls that continue to use `AuthenticationState.getCurrentPage()` do NOT need to reset the state. One of them clears the state itself, but the rest clearly do not need or want this.

The rest call now `AuthenticationState.nextRedirectOrFinish()`, which does a similar thing, but returns a `Redirect` instead of a page. The returned redirect may then be thrown, which most usages did.

This change fixes the  problems with LBA and the VPN (and other modules), brings back the "redirect to the original URL on logon" feature that was lost when the code that cleared that session was added. It is also now once again compatible with the VPN in OAuth mode (web), and the other modes continue to work correctly as before.

In general, it's also now a bit easier to follow, and is net reduction in code.

This does NOT address the use of `maybeAttachSession()`. That was taking it too far for now, as would required other framework tweaks.