Chris had a problem with support server that suddenly changed the mode it was running in (embedded mongo vs external). Being as no upgrade had taken place,  this appears to be the most likely culprit.
Application properties was ...

 * First it always loads database.properties, but the log statement says it was loading jadaptive.properties (which looks like it was the intention).
 * It then loads all .properties files in conf.d, in the order returned by OS. *Which is unpredictable*.
 * It would then load database.properties again, because *.properties covers it.
 * This meant that if any properties override others in property files, one startup of the server might result in one set of properties being used, and a 2nd startup might result in a different set.

I have changed this to be inspired by unix config files philosophy, sort the files you load first. So a certain order can be forced by prefixing the filenames with a 2 digit number. It only does exach file once in the conf.d directory.

This shouldn't have any bad effects on any properly configured server, the order will just be predictable. It *might* highlight an issue on a mis-configured one, such as was the support server.