Fixes base64 image encoding for all products in one ago. Only the default logos and gabble were currently working.
This removes all hard coded base64 encoded images, and instead will dynamically encode the image resources themselves to base64 (and cache). This is only done if needed (i.e. for embedding in emails). That is decided by the `ImageURIFormat` parameter passed when requesting the logo or the favicon from `ProductService`.

All of this is an improvement because ..

 * Improves page load times. Base64 is now only used when its needed. HTML pages (for browser) should not have been using base64 encoding images because this defeats browser caching, HTML page content  is never cached.
 * Easier maintainability. Not only would a developer have to know the existence of base64 encoded strings, they would have to keep the strings in sync with the actual images themselves.
 * Avoid massive single line images strings in Java source files, which can mess up editors.